Actualización

+++ buscar_estacionamiento/models.py
@@ -32,7 +32,7 @@ class Estacionamiento(models.Model):
     direccion = models.CharField(max_length=200)
     dueño = models.ForeignKey(Dueno, on_delete=models.CASCADE)
     comuna = models.ForeignKey(Comuna, on_delete=models.CASCADE)
-    costo_por_hora = models.DecimalField(max_digits=5, decimal_places=2, default=0.0)
+    costo_por_hora = models.IntegerField(default=0)
 
 
 class Arrendamiento(models.Model):
